+++ src/main/kotlin/io/openfuture/chain/core/exception/InsufficientTransactionsException.kt
@@ -1,0 +1,3 @@
+package io.openfuture.chain.core.exception
+
+class InsufficientTransactionsException: RuntimeException()