+++ src/main/kotlin/io/openfuture/chain/core/service/block/DefaultMainBlockService.kt
@@ -2,6 +2,7 @@ package io.openfuture.chain.core.service.block
 
 import io.openfuture.chain.consensus.property.ConsensusProperties
 import io.openfuture.chain.core.component.NodeKeyHolder
+import io.openfuture.chain.core.exception.InsufficientTransactionsException
 import io.openfuture.chain.core.model.entity.block.MainBlock
 import io.openfuture.chain.core.model.entity.block.payload.MainBlockPayload
 import io.openfuture.chain.core.repository.MainBlockRepository
@@ -12,7 +13,6 @@ import io.openfuture.chain.crypto.util.SignatureUtils
 import io.openfuture.chain.network.component.node.NodeClock
 import io.openfuture.chain.network.message.consensus.PendingBlockMessage
 import io.openfuture.chain.network.message.core.MainBlockMessage
-import io.openfuture.chain.network.service.NetworkApiService
 import org.bouncycastle.pqc.math.linearalgebra.ByteUtils
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
@@ -20,15 +20,16 @@ import org.springframework.transaction.annotation.Transactional
 @Service
 class DefaultMainBlockService(
     blockService: BlockService,
-    private val repository: MainBlockRepository,
+    repository: MainBlockRepository,
+    walletService: WalletService,
+    delegateService: DelegateService,
     private val clock: NodeClock,
     private val keyHolder: NodeKeyHolder,
     private val voteTransactionService: VoteTransactionService,
     private val delegateTransactionService: DelegateTransactionService,
     private val transferTransactionService: TransferTransactionService,
-    private val consensusProperties: ConsensusProperties,
-    private val networkService: NetworkApiService
-) : BaseBlockService(blockService), MainBlockService {
+    private val consensusProperties: ConsensusProperties
+) : BaseBlockService<MainBlock>(repository, blockService, walletService, delegateService), MainBlockService {
 
     @Transactional(readOnly = true)
     override fun create(): PendingBlockMessage {
@@ -63,14 +64,14 @@ class DefaultMainBlockService(
 
         val block = MainBlock.of(message)
         if (!isValid(block, message.getAllTransactions())) {
+            //TODO call second synchronization
             return
         }
 
-        val savedBlock = repository.save(block)
+        val savedBlock= super.save(block)
         message.voteTransactions.forEach { voteTransactionService.toBlock(it, savedBlock) }
         message.delegateTransactions.forEach { delegateTransactionService.toBlock(it, savedBlock) }
         message.transferTransactions.forEach { transferTransactionService.toBlock(it, savedBlock) }
-        networkService.broadcast(message)
     }
 
     @Transactional
@@ -83,8 +84,7 @@ class DefaultMainBlockService(
         if (!isValid(block, message.getAllTransactions().map { it.hash })) {
             return
         }
-
-        val savedBlock = repository.save(block)
+        val savedBlock = super.save(block)
         message.voteTransactions.forEach { voteTransactionService.synchronize(it, savedBlock) }
         message.delegateTransactions.forEach { delegateTransactionService.synchronize(it, savedBlock) }
         message.transferTransactions.forEach { transferTransactionService.synchronize(it, savedBlock) }
@@ -99,7 +99,7 @@ class DefaultMainBlockService(
 
     private fun calculateMerkleRoot(transactions: List<String>): String {
         if (transactions.isEmpty()) {
-            throw IllegalArgumentException("Transactions must not be empty!")
+            throw InsufficientTransactionsException()
         }
 
         if (transactions.size == 1) {