+++ src/main/kotlin/io/openfuture/chain/core/service/transaction/BaseTransactionService.kt
@@ -50,9 +50,15 @@ abstract class BaseTransactionService<T : Transaction, U : UnconfirmedTransactio
     }
 
     protected fun confirmProcess(utx: U, tx: T): T {
-        updateBalanceByFee(utx)
         unconfirmedRepository.delete(utx)
+        updateBalanceByFee(tx)
         return repository.save(tx)
+    }
+
+    protected fun isExists(hash: String): Boolean {
+        val persistUtx = unconfirmedRepository.findOneByHash(hash)
+        val persistTx = repository.findOneByHash(hash)
+        return null != persistUtx || null != persistTx
     }
 
     open fun isValid(utx: U): Boolean = this.isValidBase(utx)
@@ -67,7 +73,7 @@ abstract class BaseTransactionService<T : Transaction, U : UnconfirmedTransactio
         return isValidAddress(tx.senderAddress, tx.senderPublicKey)
             && isValidFee(tx.senderAddress, tx.fee)
             && isValidHash(tx.timestamp, tx.fee, tx.senderAddress, tx.getPayload(), tx.hash)
-            && isValidaSignature(tx.hash, tx.senderSignature, tx.senderPublicKey)
+            && isValidSignature(tx.hash, tx.senderSignature, tx.senderPublicKey)
     }
 
     private fun isValidAddress(senderAddress: String, senderPublicKey: String): Boolean {
@@ -85,7 +91,7 @@ abstract class BaseTransactionService<T : Transaction, U : UnconfirmedTransactio
         return TransactionUtils.generateHash(timestamp, fee, senderAddress, payload) == hash
     }
 
-    private fun isValidaSignature(hash: String, signature: String, publicKey: String): Boolean {
+    private fun isValidSignature(hash: String, signature: String, publicKey: String): Boolean {
         return SignatureUtils.verify(ByteUtils.fromHexString(hash), signature, ByteUtils.fromHexString(publicKey))
     }